start implemeting connection handling
